Fix logic to create ops update and final report
